avm2: Move toplevel functions to separate file
