+++ core/src/avm2/globals/toplevel.rs
@@ -1,0 +1,55 @@
+//! Global scope built-ins
+
+use crate::avm2::activation::Activation;
+use crate::avm2::object::Object;
+use crate::avm2::value::Value;
+use crate::avm2::Error;
+use crate::string::WStr;
+
+pub fn trace<'gc>(
+    activation: &mut Activation<'_, 'gc, '_>,
+    _this: Option<Object<'gc>>,
+    args: &[Value<'gc>],
+) -> Result<Value<'gc>, Error> {
+    match args {
+        [] => activation.context.avm_trace(""),
+        [arg] => {
+            let msg = arg.coerce_to_string(activation)?;
+            activation.context.avm_trace(&msg.to_utf8_lossy());
+        }
+        args => {
+            let strings = args
+                .iter()
+                .map(|a| a.coerce_to_string(activation))
+                .collect::<Result<Vec<_>, _>>()?;
+            let msg = crate::string::join(&strings, &WStr::from_units(b" "));
+            activation.context.avm_trace(&msg.to_utf8_lossy());
+        }
+    }
+
+    Ok(Value::Undefined)
+}
+
+pub fn is_finite<'gc>(
+    activation: &mut Activation<'_, 'gc, '_>,
+    _this: Option<Object<'gc>>,
+    args: &[Value<'gc>],
+) -> Result<Value<'gc>, Error> {
+    if let Some(val) = args.get(0) {
+        Ok(val.coerce_to_number(activation)?.is_finite().into())
+    } else {
+        Ok(false.into())
+    }
+}
+
+pub fn is_nan<'gc>(
+    activation: &mut Activation<'_, 'gc, '_>,
+    _this: Option<Object<'gc>>,
+    args: &[Value<'gc>],
+) -> Result<Value<'gc>, Error> {
+    if let Some(val) = args.get(0) {
+        Ok(val.coerce_to_number(activation)?.is_nan().into())
+    } else {
+        Ok(true.into())
+    }
+}
